requirements.txt
onnxruntime-gpu>=1.15.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.66.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
tensorrt>=8.5.0
torch>=2.0.0
#cupy-cuda12x